extra/ai-agent-backend/requirements.txt
aws-wsgi
boto3
pypdf
python-docx
flask
openai
pydantic
pydantic-core
pandas
psycopg2-binary
python-dotenv
requests
sqlparse
flask-cors